+++ tests/SessionNotFoundTest.php
@@ -18,8 +18,8 @@
 namespace Colopl\Spanner\Tests;
 
 use Colopl\Spanner\Connection;
-use Colopl\Spanner\Session\CacheSessionPool;
 use Google\Cloud\Core\Exception\NotFoundException;
+use Google\Cloud\Spanner\Session\CacheSessionPool;
 use Illuminate\Database\QueryException;
 use Symfony\Component\Cache\Adapter\ArrayAdapter;