+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Added
 Fixed
 - Fixed bug where running `Connection::statement` with DDLs was not logging and was not triggering events. (#86)
 
+Changed
+- Use google-cloud-php's CacheSessionPool since the [concerned bug](https://github.com/googleapis/google-cloud-php/issues/5567) has been fixed in [v1.53](https://github.com/googleapis/google-cloud-php-spanner/releases/tag/v1.58.2). (#90)
+
 # v5.0.0
 
 updated composer.json to only support laravel 10